Output slack usernames instead of full names
The output of this script is pasted into a Slack channel, so it makes sense to
have the output be the Slack username instead of the full name.

We do this by adding a new environment variable, EMAIL_TO_SLACK_MAP, which is
a JSON object mapping email addresses to Slack usernames. This is then parsed
and used to map the email addresses returned by the Opsgenie API to Slack
usernames.